moved connections models and views to a different package
